requirements.txt
# Azure Speech SDK
azure-cognitiveservices-speech>=1.34.0

# Audio processing
numpy>=1.24.0
sounddevice>=0.4.6
soxr>=0.3.7
webrtcvad>=2.0.10

# Configuration
pyyaml>=6.0
python-dotenv>=1.0.0
pydantic>=2.0.0

# Input handling
keyboard>=0.13.5

# Voicemeeter control
voicemeeterlib>=2.5.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
